+++ Layers.js
@@ -451,6 +451,146 @@ export function proceed(/* arguments */) {
   }
 };
 
+/* 
+ * Layer Class
+ */
+export class Layer {
+  constructor (name, context) {
+    this._name = name;
+    this._context = context;
+    // this._layeredFunctionsList = {};
+  }
+  
+  // Accessing
+  get name () {
+    return this._name;
+  }
+  fullName () {
+    return '' + this._context + '.' + this._name;
+  }
+  layeredObjects () {
+    return Object.getOwnPropertyNames(this)
+      .map(ea => this[ea] && this[ea]._layered_object)
+      .filter(ea => ea); // filters falsy things
+  }
+  // TODO: doesn't differentiate between functions and classes - necessary?
+  layeredClasses () {
+    return this.layeredObjects().map(ea => ea.constructor);
+  }
+  
+  // Removing
+  remove () {
+    // Deletes the LayerClass, but keeps the layered Functions.
+    if (this.isGlobal()) {
+      this.beNotGlobal();
+    }
+    var context = this._context;
+    if (typeof context !== 'undefined')
+      delete context[this.name];
+  }
+  uninstall () {
+    // Uninstalls just this Layer.
+    // functions that are layered by other Layers will not be reset.
+    var layer = this;
+    this.layeredObjects().forEach(
+      function(eachLayeredObj) {
+        // var layerIdx = typeof eachLayeredObj.activeLayers === 'function'
+        //     ? eachLayeredObj.activeLayers().indexOf(layer) : -1;
+        
+        // #Special Lively Webwerkstatt code.... General Case? #Jens
+        // #TODO if we have of gloabal list of all layers... we can look there
+        
+        // Properties.own(layer._layeredFunctionsList[eachLayeredObj]).each(
+        //   function(eachLayeredFunc) {
+        //     var newerLayer = eachLayeredObj.activeLayers().find(
+        //       function(eachOtherLayer) {
+        //         var eachOtherLayerIdx
+        //             = eachLayeredObj.activeLayers().indexOf(eachOtherLayer);
+        //         var isNewer = (eachOtherLayerIdx !== -1)
+        //             && (eachOtherLayerIdx < layerIdx);
+        //         return isNewer &&
+        //             eachOtherLayer._layeredFunctionsList[eachLayeredObj][eachLayeredFunc];
+        //       });
+        //       if (!newerLayer) {
+        //         makeFunctionLayerUnaware(eachLayeredObj, eachLayeredFunc);
+        //       }
+        //   });
+      });
+      this.remove();
+  }
+  
+  // Layer installation
+  layerClass (classObject, methods) {
+    layerClass(this, classObj, methods);
+    return this;
+  }
+  layerObject (obj, methods) {
+    layerObject(this, classObj, methods);
+    return this;
+  }
+  refineClass (classObj, methods) {
+    layerClass(this, classObj, methods);
+    return this;
+  }
+  refineObject (obj, methods) {
+    layerObject(this, obj, methods);
+    return this;
+  }
+  unrefineObject (obj) {
+    var id = obj._layer_object_id;
+    if (id !== undefined) {
+      delete this[id];
+    }
+  }
+  unrefineClass (classObj) {
+    this.unrefineObject(classObj.prototype);
+  }
+  
+  // Layer activation
+  beGlobal () {
+    enableLayer(this);
+    return this;
+  }
+  beNotGlobal () {
+    disableLayer(this);
+    return this;
+  }
+  hide () {
+    // Hidden Layers do not appear when evaluating the sourcecode of a function
+    // TODO: this function has to be called BEFORE the layer refines any class,
+    // due to problems in unrefining classes.
+    this.isHidden = true;
+    return this;
+  }
+  
+  // Testing
+  isLayer() {
+    return true;
+  }
+  isGlobal () {
+    return GlobalLayers.indexOf(this) !== -1;
+  }
+  
+  // Debugging
+  toString () {
+    return this.name;
+  }
+  
+  // Deprecated serialization
+  toLiteral () {
+    if (!this.name) {
+      console.warn("Layer: Can not serialize without a name!");
+    }
+    return { name: this.name };
+  }
+  
+  // Deserialization
+  fromLiteral (literal) {
+    // console.log("Deserializing Layer activation from: " + literal.name);
+    return create(literal.name, false);
+  }
+}
+
 /*
  * Example implementation of a layerable object
  */
